editoast: move ModelV2 macro code generation into a dedicated module
